adk/pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "dotenv>=0.9.9",
    "google-adk==1.5.0",
    "litellm>=1.76.1",
    "mcp[cli]>=1.13.1",
    "openai>=1.104.2",
    "requests>=2.32.5",
]
